refactor: remove shared folder and move contents of shared folder to app directory
